==> Code/1-tidy_data.Rmd <==
---
title: "Tidy and subset lens data"
author: "Amartya Mitra, Molly Womack, Katie Thomas"
date: '2022-07-20'
output:
  html_document:
    code_fold: show
    theme: flatly
    toc: yes
    toc_float: yes
---

<style type="text/css">

body{ /* Normal  */
      font-size: 17px;
  }
  
</style>

```{r setup, include = FALSE}

# Markdown settings
knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE, fig.height = 6, fig.width = 8, attr.output='style="max-height: 400px;"') 

# Load packages
library(ape)
library(geiger)
library(kableExtra)
library(tidyverse)
```

# Raw data

## Raw lens data

The raw data file contains one row for **each lens** sampled. For most specimens, there are two rows of data: one for the right lens and one for the left lens. However, for a few specimens, we repeated measurements on the lenses independently to quantify error, and these are included as separate rows. The raw data file also includes measurements from four artificial objects created in Blender for reference measurements. These are all coded as life stage "artificial", and can be removed or extracted that way.

```{r}
# Load raw data
lens_data <- read.csv("../Data/Raw/Mitra_et_al_data_revised.csv", header = TRUE, na.strings=c("", "NA", " "))
```

## Raw phylogeny

The phylogeny data is the consensus tree from Jetz and Pyron (2019).

```{r}
#Import phylogeny from Jetz and Pyron 2019
tree_orig <- ape::read.tree(file = "../Data/Raw/amph_shl_new_Consensus_7238.tre") #reads tree
```

# Tidy lens data

Here, we remove an individual (BMNH1977.1533) from the dataset we will use for analyses. While our paper was in review, a reviewer asked us to take a careful look at our data for an adult Xenopus victorianus. To address this, we downloaded a scan of a different individual from Morphosource (CAS:SUA:18268) and used data from this instead of from our original scan (BMNH1977.1533). 

```{r}

#tidy lens data
lens_data_tidy <- lens_data %>%
  filter(specimen_id != "BMNH1977.1533") %>%
  mutate(activity_period = recode(activity_period, nocturnal = "nocturnal", diurnal = "non-nocturnal", both = "non-nocturnal"))

#export tidy data
write.csv(lens_data_tidy, file = "../Data/Tidy/lens_data_tidy.csv", row.names = FALSE)
```


# Subset lens data

## Mean lens shape by specimen

The following subset includes one row for **each specimen** measured. We created this subset by removing repeated measurements of the same lens, and then calculating the mean of the left and right lens for each specimen. This subset also includes artificial reference blender shapes. 

```{r specimen-means}

#Calculate mean lens parameters for each specimen
specimen_means <- lens_data_tidy %>%
  mutate(genus_species = as.factor(paste(genus, species, sep = "_"))) %>% # creates a genus_species column
  filter(life_stage != "repeat") %>%
  mutate_if(is.character, as.factor) %>% 
  group_by(specimen_id, genus_species, family, life_stage, adult_ecology, activity_period, scan_source, stain) %>%
  summarise(mean_flatness = mean(flatness, na.rm = TRUE),
            mean_anisotropy = mean(anisotropy, na.rm = TRUE),
            mean_elongation = mean(elongation, na.rm = TRUE),
            mean_sphericity = mean(sphericity, na.rm = TRUE),
            mean_svl = mean(sv_length_mm, na.rm = TRUE),
            n = n()) %>%
  ungroup() 

#convert NaN values to NA 
specimen_means[specimen_means == "NaN"] <- NA

#export specimen means data
write.csv(specimen_means, file = "../Data/Tidy/specimen_means.csv", row.names = FALSE)
```

## Summary data on specimens

Here, we report summary data on specimen sampling (after removing the 4 artificial reference objects created in Blender). Note that the specimen ID "MS2_top" refers to both a tadpole and an adult specimen, and is a duplicated element (thus there is one less factor level for specimen ID than there are actual unique specimens).

```{r}
#remove artificial objects from dataset
specimen_means2 <- specimen_means %>%
  filter(life_stage != "artificial") %>%
  droplevels()

#save data on sampling
n_specimen <- nrow(specimen_means2) #total number of specimens sampled
n_adult_specimens <- nrow(specimen_means2 %>% filter(life_stage=="adult"))
n_tad_specimens <- nrow(specimen_means2 %>% filter(life_stage=="tadpole"))
n_family <- length(levels(specimen_means2$family))
n_species <- length(levels(specimen_means2$genus_species))
```

Number of specimens sampled total: `r n_specimen`

Number of adult specimens sampled: `r n_adult_specimens`

Number of tadpole specimens sampled: `r n_tad_specimens`

Number of families sampled: `r n_family`

Number of species sampled: `r n_species`

Number of specimens sampled by source: 

```{r}
#number of specimens by source
specimen_means2 %>%
  count(scan_source)

#number of specimens by source and life stage
specimen_means2 %>%
  group_by(life_stage) %>%
  count(scan_source)
```

Number of specimens sampled by staining:

```{r}
#number of specimens by source
specimen_means2 %>%
  group_by(life_stage) %>%
  count(stain)
```

## Mean lens shape by species and life stage

### Life stages in different rows

The following creates a dataframe with **one row per life stage for a species**. Mean shape data for **tadpoles** and for **adults** are in separate rows and any duplicated life stages within species are averaged. This subset contains anurans only (no artificial reference shapes).

```{r species-stage-means}

species_stage_means <- specimen_means %>%
  filter(life_stage != "artificial") %>%
  group_by(genus_species, family, life_stage, adult_ecology, activity_period) %>%
  summarise(mean_flatness = mean(mean_flatness, na.rm = TRUE),
            mean_anisotropy = mean(mean_anisotropy, na.rm = TRUE),
            mean_elongation = mean(mean_elongation, na.rm = TRUE),
            mean_sphericity = mean(mean_sphericity, na.rm = TRUE),
            mean_svl = mean(mean_svl, na.rm = TRUE),
            n_specimens = n()) %>%
    mutate(mean_svl = na_if(mean_svl, "NaN"))


#find species in dataset that don't exist as phylogeny tip label
missing <- as.vector(which(!species_stage_means$genus_species %in% as.vector(tree_orig$tip.label)))
(as.data.frame(species_stage_means))[missing, 1]

#add synonym and substitutions for phylogeny into dataframe for the 5 species that aren't matching (phylo_species column)
species_stage_means$phylo_species <- as.character(species_stage_means$genus_species)
species_stage_means[species_stage_means$phylo_species=="Boana_geographica", "phylo_species"] <- "Hypsiboas_geographicus"
species_stage_means[species_stage_means$phylo_species=="Chiasmocleis_hudsoni", "phylo_species"] <- "Syncope_hudsoni"
species_stage_means[species_stage_means$phylo_species=="Elachistocleis_nigrogularis", "phylo_species"] <- "Elachistocleis_ovalis"
species_stage_means[species_stage_means$phylo_species=="Indosylvirana_temporalis", "phylo_species"] <- "Hylarana_temporalis"
species_stage_means[species_stage_means$phylo_species=="Rhombophryne_nilevina", "phylo_species"] <- "Rhombophryne_alluaudi"

#remove extra levels
species_stage_means <- species_stage_means %>%
  droplevels()

#export species means by life stage
write.csv(species_stage_means, file = "../Data/Tidy/species_stage_means.csv", row.names = FALSE)
```

### Life stages in different columns

The following creates a dataframe with **one row per species**, with tadpole and adult data in different columns.

```{r species_means-wide}

#make separate dataframes for adults and tadpoles
tads_mean <- filter(species_stage_means, life_stage == "tadpole") %>% droplevels()
adults_mean <- filter(species_stage_means, life_stage == "adult") %>% droplevels()

#merge tadpole and adult dataframes by species (all data)
species_means <- full_join(tads_mean, adults_mean, 
                           by = "genus_species", 
                           suffix = c(".tads", ".adults")) %>%
  rename(adult_ecology = adult_ecology.adults) %>%
  rename(activity_period = activity_period.adults) %>%
  mutate(phylo_species = ifelse(is.na(phylo_species.adults), as.character(phylo_species.tads), as.character(phylo_species.adults))) %>%
  mutate(family = ifelse(is.na(family.adults), as.character(family.tads), as.character(family.adults))) %>%
  select(genus_species, family, mean_flatness.tads, mean_anisotropy.tads, mean_elongation.tads, mean_sphericity.tads, mean_svl.tads, n_specimens.tads, mean_flatness.adults, mean_anisotropy.adults, mean_elongation.adults, mean_sphericity.adults, mean_svl.adults, n_specimens.adults, adult_ecology, activity_period, phylo_species)

#export species means
write.csv(species_means, file = "../Data/Tidy/species_means.csv", row.names = FALSE)
```

## Subset with adult and tadpole matched data

The following creates a dataframe with **one row per species**, with tadpole and adult data in different columns, and subsets to only species with both tadpole and adult sampling. 

```{r}
#merge only matched tadpole and adult data within species
species_matched <- inner_join(tads_mean, adults_mean,by = c("genus_species", "family"), suffix = c(".tads", ".adults")) %>% 
  droplevels() %>%
  rename(adult_ecology = adult_ecology.adults) %>%
  rename(activity_period = activity_period.adults) %>%
  mutate(mean_svl.tads = na_if(mean_svl.tads, "NaN")) %>%
  select(genus_species, family, mean_flatness.tads, mean_anisotropy.tads, mean_elongation.tads, mean_sphericity.tads, mean_svl.tads, n_specimens.tads, mean_flatness.adults, mean_anisotropy.adults, mean_elongation.adults, mean_sphericity.adults, mean_svl.adults, n_specimens.adults, adult_ecology, activity_period)

#export tad/adult matched data
write.csv(species_matched, file = "../Data/Tidy/species_matched.csv", row.names = FALSE)
```

## Sampling by life stage

```{r}
#sampling for adults and tadpoles
n_tadpole_species <- length(levels(as.factor(tads_mean$genus_species)))
n_adult_species <- length(levels(as.factor(adults_mean$genus_species)))
n_matched_species <- length(levels(as.factor(species_matched$genus_species)))
n_matched_families <- length(levels(as.factor(species_matched$family)))
```

**Species sampling by life stage:**

Number of species with adult data: `r n_adult_species`

Number of species with tadpole data: `r n_tadpole_species`

Number of species with both tadpole and adult data (matched): `r n_matched_species`

Number of families with tadpole/adult matched data: `r n_matched_families`


## Species with eye size data

This dataframe pulls out the specimens that we were able to get 2D eye size measurements for in Avizo. The mean eye measurements are calculated for each **specimen**. Then, since there were few/no tadpoles with data, the **adults** are subset out for further analyses. 

```{r}

#create dataframe for adult eye and lens subset
adult_eye <- lens_data_tidy %>%
  mutate(genus_species = as.factor(paste(genus, species, sep = "_"))) %>% 
  mutate_if(is.character, as.factor) %>% 
  filter(life_stage == "adult") %>%
  #remove rows without eye size data
  filter(!is.na(eye_diameter_mm)) %>% 
  #remove rows without lens size data
  filter(!is.na(lens_diameter_mm)) %>% 
  group_by(genus_species, life_stage, adult_ecology, activity_period) %>%
  summarise(mean.eye.diameter = mean(eye_diameter_mm, na.rm = TRUE),
            mean.lens.diameter = mean(lens_diameter_mm, na.rm = TRUE),
            n = n()) %>%
  mutate(lens_relative_to_eye = mean.lens.diameter/mean.eye.diameter) %>%
  droplevels()

#export eye and lens size dataset
write.csv(adult_eye, file = "../Data/Tidy/adult_eye.csv", row.names = FALSE)
```

Species with 2D adult eye & lens size data:  `r nrow(adult_eye)`


# Prune phylogeny to match species with data

Here, we prune the phylogeny to the 126 species in our dataset, and then relabel the tip labels to match our names for the species. The phylogeny we are using has two species that differ in name from the species in our dataset (synonyms), and one species that is not represented. Specifically, in the phylogeny we replace _Hypsiboas geographicus_ with synonym _Boana geographica_, _Syncope hudsoni_ with synonym _Chiasmocleis hudsoni_, and _Hylarana temporalis_ with synonym _Indosylvirana temporalis_. We replace _Rhombophryne_alluaudi_ with substitute _Rhombophryne_nilevina_ and _Elachistocleis ovalis_ with substitute _Elachistocleis nigrorgularis_.

```{r, fig.height=12, fig.width=5}

#make new dataframe of species means
tree_data <- ungroup(species_means)

#add rownames to data
rownames(tree_data) <- tree_data$phylo_species

#check whether names in tree and dataset match
name.check(tree_orig, tree_data) -> overlap

#species in data that aren't matching tree
overlap$data_not_tree

#drop tips from tree that don't match the data
tree.pruned <- drop.tip(tree_orig, overlap$tree_not_data)
plot(tree.pruned, cex = 0.6)

#create a dataframe with column 1 with all original phylo tip labels, and column 2 with corresponding desired tip labels
subs <- tree_data[ , c("phylo_species", "genus_species")] %>%
  ungroup() %>%
  mutate(across(everything(), as.character))

#make a new tree to manipulate
tree.sub <- tree.pruned

#rename tree tip labels to match the species names in our dataset
tree.sub$tip.label <- subs[[2]][match(tree.sub$tip.label, subs[[1]])]

#check that tree tip lables match data species labels
setdiff(tree.sub$tip.label, tree_data$genus_species)

#plot final pruned, relabeled tree
plot.phylo(tree.sub, #phylogeny to plot
           type = "phylogram", #how to visualize phylogeny
           show.tip.label = TRUE, #whether to show tip labels/species names
           cex = 0.6, #text size
           no.margin = TRUE, 
           use.edge.length = TRUE,
           edge.width = 1.5, #thickness of phylogeny branches
           label.offset = 0.1) #how far away from tips to put labels

#export final tree
write.nexus(tree.sub, file = "../Data/Tidy/lens_shape_tree.nex")
```
